requirements.txt
alabaster==0.7.13
Babel==2.12.1
blinker==1.6.2
blurb==1.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
docutils==0.20.1
Flask==2.3.2
idna==3.4
imagesize==1.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
Pygments==2.15.1
PyPDF2==3.0.1
python-doc==2020.10.1
python-docs-theme==2023.5
python-docx==0.8.11
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==2.0.3
Werkzeug==2.3.6
boostrap  5.3.0
cdnjquery 3.7.0 cdn
python 3.11.4
